refactor: split out utlities into their modules and change api to use metadata
